fix changes from merge
